requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dataclasses-json==0.5.7
dnspython==2.3.0
Flask==2.3.2
Flask-Session==0.5.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.189
livereload==2.6.3
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pinecone-client==3.0.1
pycparser==2.21
pydantic==1.10.8
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tiktoken==0.4.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.2
Wand==0.6.11
Werkzeug==2.3.6
win32-setctime==1.1.0
yarl==1.9.2
